requirements.txt
# UrsaOps Python Requirements

PyYAML==5.4.1
Jinja2==3.1.2
